+++ src/main/java/fi/digitraffic/graphql/rail/to/OperatorTOConverter.java
@@ -13,7 +13,7 @@ public class OperatorTOConverter extends BaseConverter<OperatorTO> {
         return new OperatorTO(
                 tuple.get(QOperator.operator.name),
                 tuple.get(QOperator.operator.shortCode),
-                tuple.get(QOperator.operator.operatorUicCode).longValue()
+                tuple.get(QOperator.operator.operatorUicCode).intValue()
         );
     }
 }
