+++ src/main/java/fi/digitraffic/graphql/rail/to/RoutesetMessageTOConverter.java
@@ -10,7 +10,7 @@ import fi.digitraffic.graphql.rail.model.RoutesetMessageTO;
 public class RoutesetMessageTOConverter extends BaseConverter<RoutesetMessageTO> {
     public RoutesetMessageTO convert(final Tuple tuple) {
         return new RoutesetMessageTO(
-                tuple.get(QRouteset.routeset.id).longValue(),
+                tuple.get(QRouteset.routeset.id).intValue(),
                 tuple.get(QRouteset.routeset.version).toString(),
                 tuple.get(QRouteset.routeset.messageTime),
                 tuple.get(QRouteset.routeset.trainId.trainNumber),
