+++ src/main/java/fi/digitraffic/graphql/rail/to/TrackRangeTOConverter.java
@@ -10,7 +10,7 @@ import fi.digitraffic.graphql.rail.model.TrackRangeTO;
 public class TrackRangeTOConverter {
     public TrackRangeTO convert(final Tuple tuple) {
         return new TrackRangeTO(
-                tuple.get(QTrackRange.trackRange.trackSectionId),
+                tuple.get(QTrackRange.trackRange.trackSectionId).intValue(),
                 tuple.get(QTrackRange.trackRange.startTrack),
                 tuple.get(QTrackRange.trackRange.startKilometres),
                 tuple.get(QTrackRange.trackRange.startMetres),
