+++ src/main/java/fi/digitraffic/graphql/rail/to/TrainTrackingTOConverter.java
@@ -12,7 +12,7 @@ import fi.digitraffic.graphql.rail.model.TrainTrackingMessageTypeTO;
 public class TrainTrackingTOConverter {
     public TrainTrackingMessageTO convert(final Tuple tuple) {
         return new TrainTrackingMessageTO(
-                tuple.get(QTrainTrackingMessage.trainTrackingMessage.id).longValue(),
+                tuple.get(QTrainTrackingMessage.trainTrackingMessage.id).intValue(),
                 tuple.get(QTrainTrackingMessage.trainTrackingMessage.trainId.trainNumber),
                 tuple.get(QTrainTrackingMessage.trainTrackingMessage.trainId.virtualDepartureDate),
                 tuple.get(QTrainTrackingMessage.trainTrackingMessage.stationShortCode),
