+++ src/main/java/fi/digitraffic/graphql/rail/to/TrainTypeTOConverter.java
@@ -11,9 +11,9 @@ public class TrainTypeTOConverter extends BaseConverter<TrainTypeTO> {
     @Override
     public TrainTypeTO convert(final Tuple tuple) {
         return new TrainTypeTO(
-                tuple.get(QTrainType.trainType.id).longValue(),
+                tuple.get(QTrainType.trainType.id).intValue(),
                 tuple.get(QTrainType.trainType.name),
-                tuple.get(QTrainType.trainType.trainCategoryId).longValue()
+                tuple.get(QTrainType.trainType.trainCategoryId).intValue()
                 , null
         );
     }
